Unify error handling between commands
